[Refactor] Register 구조 재정리(#61)
