+++ src/routes/Register/Register.css
@@ -1,6 +1,6 @@
 
 .register-main {
-    background-image: url('../assets/images/login_background.png');
+    background-image: url('../../assets/images/login_background.png');
     background-size: cover;
     background-position: center;
     background-repeat: no-repeat;
